+++ src/components/Title/TitleInput/SubtitleTextInput/SubtitleTextInput.style.ts
@@ -1,0 +1,1 @@
+import styled from "styled-components";
